marketplace/requirements.txt
grpcio-tools==1.56.0
flask==2.0.1
jinja2==3.0.1
pytest==6.2.5
